Fix decision context CI integration
