+++ internal/agent/contracts/schemas.gen.go
@@ -31,6 +31,6 @@ const (
 	DocsReadResultSchemaJSON = `{"additionalProperties":false,"properties":{"content":{"type":"string"},"id":{"type":"string"},"lineEnd":{"type":"integer"},"lineStart":{"type":"integer"},"nextOffset":{"type":"integer"},"path":{"type":"string"},"title":{"type":"string"},"totalLines":{"type":"integer"},"truncated":{"type":"boolean"},"url":{"type":"string"}},"required":["content","id","lineEnd","lineStart","path","title","totalLines","truncated","url"],"type":"object"}`
 	DocsSearchInputSchemaJSON = `{"additionalProperties":false,"properties":{"cursor":{"description":"Opaque nextCursor from a previous call with the same query and path.","minLength":1,"type":"string"},"limit":{"description":"Maximum matches to return; defaults to 8.","maximum":25,"minimum":1,"type":"integer"},"path":{"description":"Optional documentation path prefix, such as guides/build, api, cli, or visuals.","type":"string"},"query":{"description":"Words or phrases to find in LeapView documentation.","minLength":1,"type":"string"}},"required":["query"],"type":"object"}`
 	DocsSearchResultSchemaJSON = `{"additionalProperties":false,"properties":{"count":{"minimum":0,"type":"integer"},"hasMore":{"type":"boolean"},"matches":{"items":{"additionalProperties":false,"properties":{"excerpt":{"type":"string"},"id":{"type":"string"},"path":{"type":"string"},"summary":{"type":"string"},"title":{"type":"string"},"url":{"type":"string"}},"required":["excerpt","id","path","summary","title","url"],"type":"object"},"type":"array"},"nextCursor":{"type":"string"},"path":{"type":"string"},"query":{"type":"string"}},"required":["count","hasMore","matches","query"],"type":"object"}`
-	QueryVisualInputSchemaJSON = `{"additionalProperties":false,"properties":{"filters":{"description":"Governed semantic filters applied before visualization. Inline data and arbitrary expressions are not accepted.","items":{"additionalProperties":false,"properties":{"control":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"options":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"dependsOn":{"items":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"type":"array"},"includeNull":{"type":"boolean"},"limit":{"type":"integer"},"type":{"enum":["distinct"],"type":"string"}},"required":["dataset","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["static"],"type":"string"},"values":{"items":{"additionalProperties":false,"properties":{"label":{"type":"string"},"value":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["string"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["boolean"],"type":"string"},"value":{"type":"boolean"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["integer"],"type":"string"},"value":{"description":"Precision-safe integer representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["decimal"],"type":"string"},"value":{"description":"Canonical decimal representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["date"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["timestamp"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"}]}},"required":["label","value"],"type":"object"},"type":"array"}},"required":["type","values"],"type":"object"}]},"type":{"enum":["singleSelect"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"maxSelectedValues":{"type":"integer"},"options":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"dependsOn":{"items":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"type":"array"},"includeNull":{"type":"boolean"},"limit":{"type":"integer"},"type":{"enum":["distinct"],"type":"string"}},"required":["dataset","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["static"],"type":"string"},"values":{"items":{"additionalProperties":false,"properties":{"label":{"type":"string"},"value":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["string"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["boolean"],"type":"string"},"value":{"type":"boolean"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["integer"],"type":"string"},"value":{"description":"Precision-safe integer representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["decimal"],"type":"string"},"value":{"description":"Canonical decimal representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["date"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["timestamp"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"}]}},"required":["label","value"],"type":"object"},"type":"array"}},"required":["type","values"],"type":"object"}]},"type":{"enum":["multiSelect"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["text"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["numericRange"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["dateRange"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["relativePeriod"],"type":"string"}},"required":["type"],"type":"object"}]},"default":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["unfiltered"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"operator":{"enum":["isNull","isNotNull","in","notIn","equals","notEquals","contains","notContains","startsWith","endsWith","greaterThan","greaterThanOrEqual","lessThan","lessThanOrEqual"],"type":"string"},"type":{"enum":["nullCheck"],"type":"string"}},"required":["operator","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"operator":{"enum":["isNull","isNotNull","in","notIn","equals","notEquals","contains","notContains","startsWith","endsWith","greaterThan","greaterThanOrEqual","lessThan","lessThanOrEqual"],"type":"string"},"type":{"enum":["set"],"type":"string"},"values":{"items":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["string"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["boolean"],"type":"string"},"value":{"type":"boolean"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["integer"],"type":"string"},"value":{"description":"Precision-safe integer representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["decimal"],"type":"string"},"value":{"description":"Canonical decimal representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["date"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["timestamp"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"}]},"type":"array"}},"required":["operator","type","values"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"operator":{"enum":["isNull","isNotNull","in","notIn","equals","notEquals","contains","notContains","startsWith","endsWith","greaterThan","greaterThanOrEqual","lessThan","lessThanOrEqual"],"type":"string"},"type":{"enum":["comparison"],"type":"string"},"value":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["string"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["boolean"],"type":"string"},"value":{"type":"boolean"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["integer"],"type":"string"},"value":{"description":"Precision-safe integer representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["decimal"],"type":"string"},"value":{"description":"Canonical decimal representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["date"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["timestamp"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"}]}},"required":["operator","type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"lower":{"additionalProperties":false,"properties":{"inclusive":{"type":"boolean"},"value":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["string"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["boolean"],"type":"string"},"value":{"type":"boolean"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["integer"],"type":"string"},"value":{"description":"Precision-safe integer representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["decimal"],"type":"string"},"value":{"description":"Canonical decimal representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["date"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["timestamp"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"}]}},"required":["inclusive","value"],"type":"object"},"type":{"enum":["range"],"type":"string"},"upper":{"additionalProperties":false,"properties":{"inclusive":{"type":"boolean"},"value":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["string"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["boolean"],"type":"string"},"value":{"type":"boolean"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["integer"],"type":"string"},"value":{"description":"Precision-safe integer representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["decimal"],"type":"string"},"value":{"description":"Canonical decimal representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["date"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["timestamp"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"}]}},"required":["inclusive","value"],"type":"object"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"anchor":{"enum":["currentTime","firstAvailable","lastAvailable","fixed"],"type":"string"},"anchorValue":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["string"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["boolean"],"type":"string"},"value":{"type":"boolean"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["integer"],"type":"string"},"value":{"description":"Precision-safe integer representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["decimal"],"type":"string"},"value":{"description":"Canonical decimal representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["date"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["timestamp"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"}]},"count":{"type":"integer"},"direction":{"enum":["previous","current","next"],"type":"string"},"includeCurrent":{"type":"boolean"},"type":{"enum":["relativePeriod"],"type":"string"},"unit":{"enum":["minute","hour","day","week","month","quarter","year"],"type":"string"}},"required":["anchor","count","direction","includeCurrent","type","unit"],"type":"object"}]},"description":{"type":"string"},"dimension":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"type":"string"},"operators":{"items":{"enum":["isNull","isNotNull","in","notIn","equals","notEquals","contains","notContains","startsWith","endsWith","greaterThan","greaterThanOrEqual","lessThan","lessThanOrEqual"],"type":"string"},"type":"array"},"readerEditable":{"type":"boolean"},"required":{"type":"boolean"},"targets":{"items":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"type":"array"},"urlParameter":{"maxLength":64,"minLength":1,"pattern":"^[A-Za-z][A-Za-z0-9_-]{0,63}$","type":"string"}},"required":["control","dimension","id","label"],"type":"object"},"type":"array"},"semanticModelId":{"description":"Stable semantic_model resource ID returned by catalog_search/catalog_get.","minLength":1,"type":"string"},"visual":{"additionalProperties":false,"properties":{"accessibility":{"additionalProperties":false,"properties":{"announceChanges":{"type":"boolean"},"description":{"type":"string"},"summary":{"type":"string"},"title":{"type":"string"}},"type":"object"},"calculations":{"items":{"additionalProperties":false,"properties":{"axis":{"enum":["rows","columns","hierarchy","facets"],"type":"string"},"hidden":{"type":"boolean"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"type":"string"},"lookup":{"additionalProperties":false,"properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"value":{"type":"string"}},"required":["field","value"],"type":"object"},"offset":{"type":"integer"},"orderBy":{"items":{"additionalProperties":false,"properties":{"direction":{"enum":["asc","desc"],"type":"string"},"field":{"description":"A compiled result-frame field name, never a source member or expression.","minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["direction","field"],"type":"object"},"type":"array"},"parent":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"partitionBy":{"items":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"type":"array"},"reset":{"enum":["none","highest_parent","lowest_parent"],"type":"string"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"template":{"enum":["running_total","moving_average","difference","percentage_difference","percent_of_parent","percent_of_grand_total","rank","cumulative_contribution","lookup"],"type":"string"},"window":{"type":"integer"}},"required":["id","source","template"],"type":"object"},"type":"array"},"dataBudget":{"additionalProperties":false,"properties":{"maxRows":{"type":"integer"},"requiredCompleteness":{"enum":["complete","truncated","partial","empty"],"type":"string"}},"required":["maxRows"],"type":"object"},"datasets":{"additionalProperties":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"dimensions":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic dimension reference. Physical dataset fields are not accepted.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"dimension":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"grain":{"enum":["second","minute","hour","day","week","month","quarter","year"],"type":"string"}},"required":["dimension"],"type":"object"}]},"type":"array"},"limit":{"type":"integer"},"metrics":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic metric reference. Metric expressions are never part of this type.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"metric":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["metric"],"type":"object"}]},"type":"array"},"sort":{"items":{"additionalProperties":false,"properties":{"direction":{"enum":["asc","desc"],"type":"string"},"field":{"description":"A compiled result-frame field name, never a source member or expression.","minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["direction","field"],"type":"object"},"type":"array"},"type":{"enum":["aggregate"],"type":"string"}},"required":["dimensions","metrics","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"fields":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A physical field reference is only valid for a records query root dataset.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["field"],"type":"object"}]},"type":"array"},"limit":{"type":"integer"},"sort":{"items":{"additionalProperties":false,"properties":{"direction":{"enum":["asc","desc"],"type":"string"},"field":{"description":"A compiled result-frame field name, never a source member or expression.","minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["direction","field"],"type":"object"},"type":"array"},"type":{"enum":["records"],"type":"string"}},"required":["dataset","fields","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"columns":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic dimension reference. Physical dataset fields are not accepted.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"dimension":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"grain":{"enum":["second","minute","hour","day","week","month","quarter","year"],"type":"string"}},"required":["dimension"],"type":"object"}]},"type":"array"},"metrics":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic metric reference. Metric expressions are never part of this type.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"metric":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["metric"],"type":"object"}]},"type":"array"},"rows":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic dimension reference. Physical dataset fields are not accepted.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"dimension":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"grain":{"enum":["second","minute","hour","day","week","month","quarter","year"],"type":"string"}},"required":["dimension"],"type":"object"}]},"type":"array"},"sort":{"items":{"additionalProperties":false,"properties":{"direction":{"enum":["asc","desc"],"type":"string"},"field":{"description":"A compiled result-frame field name, never a source member or expression.","minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["direction","field"],"type":"object"},"type":"array"},"totals":{"additionalProperties":false,"properties":{"columns":{"type":"boolean"},"grand":{"type":"boolean"},"rows":{"type":"boolean"}},"type":"object"},"type":{"enum":["pivot"],"type":"string"},"window":{"additionalProperties":false,"properties":{"limit":{"type":"integer"},"offset":{"type":"integer"}},"required":["limit"],"type":"object"}},"required":["columns","metrics","rows","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"approximation":{"enum":["exact","approximate"],"type":"string"},"bins":{"type":"integer"},"domain":{"additionalProperties":false,"properties":{"maximum":{"type":"number"},"minimum":{"type":"number"}},"type":"object"},"field":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic metric reference. Metric expressions are never part of this type.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"metric":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["metric"],"type":"object"}]},"nullPolicy":{"enum":["omit","include"],"type":"string"},"type":{"enum":["histogram"],"type":"string"}},"required":["approximation","bins","field","nullPolicy","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"approximation":{"enum":["exact","approximate"],"type":"string"},"field":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic metric reference. Metric expressions are never part of this type.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"metric":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["metric"],"type":"object"}]},"group":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic dimension reference. Physical dataset fields are not accepted.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"dimension":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"grain":{"enum":["second","minute","hour","day","week","month","quarter","year"],"type":"string"}},"required":["dimension"],"type":"object"}]},"limit":{"type":"integer"},"outliers":{"enum":["omit","include"],"type":"string"},"quantiles":{"items":{"type":"number"},"type":"array"},"type":{"enum":["distribution"],"type":"string"},"whiskers":{"additionalProperties":false,"properties":{"lower":{"type":"number"},"upper":{"type":"number"}},"required":["lower","upper"],"type":"object"}},"required":["approximation","field","outliers","quantiles","type"],"type":"object"}]},"propertyNames":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"type":"object"},"description":{"type":"string"},"interactions":{"items":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"targets":{"items":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"type":"array"},"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"highlightTargets":{"items":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"type":"array"},"mappings":{"items":{"additionalProperties":false,"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"grain":{"enum":["second","minute","hour","day","week","month","quarter","year"],"type":"string"},"label":{"type":"string"},"value":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["field","value"],"type":"object"},"type":"array"},"mode":{"enum":["single","multiple"],"type":"string"},"noneTargets":{"items":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"type":"array"},"toggle":{"type":"boolean"},"type":{"enum":["selection"],"type":"string"}},"required":["mappings","mode","toggle","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"targets":{"items":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"type":"array"},"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"gestures":{"items":{"enum":["box","lasso","radius"],"type":"string"},"type":"array"},"highlightTargets":{"items":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"type":"array"},"latitude":{"additionalProperties":false,"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["field","source"],"type":"object"},"longitude":{"additionalProperties":false,"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["field","source"],"type":"object"},"noneTargets":{"items":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"type":"array"},"type":{"enum":["spatialSelection"],"type":"string"}},"required":["gestures","latitude","longitude","type"],"type":"object"}]},"type":"array"},"metadata":{"additionalProperties":false,"properties":{"description":{"additionalProperties":false,"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"fallback":{"type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"prefix":{"type":"string"},"reducer":{"enum":["first","last","minimum","maximum","mean","median"],"type":"string"},"suffix":{"type":"string"}},"required":["dataset","field"],"type":"object"},"subtitle":{"additionalProperties":false,"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"fallback":{"type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"prefix":{"type":"string"},"reducer":{"enum":["first","last","minimum","maximum","mean","median"],"type":"string"},"suffix":{"type":"string"}},"required":["dataset","field"],"type":"object"},"summary":{"additionalProperties":false,"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"fallback":{"type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"prefix":{"type":"string"},"reducer":{"enum":["first","last","minimum","maximum","mean","median"],"type":"string"},"suffix":{"type":"string"}},"required":["dataset","field"],"type":"object"},"title":{"additionalProperties":false,"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"fallback":{"type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"prefix":{"type":"string"},"reducer":{"enum":["first","last","minimum","maximum","mean","median"],"type":"string"},"suffix":{"type":"string"}},"required":["dataset","field"],"type":"object"}},"type":"object"},"presentation":{"oneOf":[{"additionalProperties":false,"allOf":[{"description":"Renderer-neutral presentation policies; renderer options are intentionally absent.","properties":{"conditionalFormatting":{"items":{"additionalProperties":false,"description":"A closed, governed conditional-format binding over compiled result names.","properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"rule":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"high":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["gradient"],"type":"string"},"low":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"maximum":{"type":"number"},"minimum":{"type":"number"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"}},"required":["high","kind","low","maximum","minimum","nullStyle"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["rules"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"rules":{"items":{"additionalProperties":false,"description":"One ordered numeric comparison in a rules-based conditional format.","properties":{"operator":{"enum":["less_than","less_or_equal","greater_than","greater_or_equal","equal","not_equal"],"type":"string"},"style":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"value":{"type":"number"}},"required":["operator","style","value"],"type":"object"},"type":"array"}},"required":["defaultStyle","kind","nullStyle","rules"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["field"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"values":{"additionalProperties":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"type":"object"}},"required":["defaultStyle","kind","nullStyle","source","values"],"type":"object"}]},"target":{"enum":["mark_fill","mark_stroke","series_color","label_foreground","visual_background","cell_foreground","cell_background","kpi_value","icon"],"type":"string"}},"required":["field","id","rule","target"],"type":"object"},"type":"array"},"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"dataZoom":{"type":"boolean"},"displayUnits":{"enum":["auto","none","thousands","millions","billions","trillions"],"type":"string"},"labelPosition":{"enum":["automatic","inside","outside","top"],"type":"string"},"labels":{"additionalProperties":false,"properties":{"density":{"enum":["hidden","automatic","dense","always"],"type":"string"},"maxCharacters":{"type":"integer"},"minimumSpacing":{"type":"integer"},"priority":{"items":{"enum":["selected","anomaly","threshold"],"type":"string"},"type":"array"},"tooltipFallback":{"type":"boolean"}},"required":["density"],"type":"object"},"legend":{"enum":["none","top","right","bottom","left"],"type":"string"},"orientation":{"enum":["horizontal","vertical"],"type":"string"},"series":{"description":"Explicit per-measure marks and axis assignments for combo visuals.","items":{"additionalProperties":false,"description":"One result-frame measure and its governed combo rendering policy.","properties":{"axis":{"description":"Axis assignment for one governed combo measure.","enum":["primary","secondary"],"type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"mark":{"description":"Marks supported by a governed multi-measure combo series.","enum":["line","area","bar","column"],"type":"string"}},"required":["axis","field","mark"],"type":"object"},"type":"array"},"showSymbols":{"type":"boolean"},"smooth":{"type":"boolean"},"stacking":{"enum":["none","normal","percent"],"type":"string"},"step":{"description":"Render line segments as discrete steps between categories.","type":"boolean"},"symbolSize":{"type":"number"},"type":{"enum":["cartesian"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"Renderer-neutral presentation policies; renderer options are intentionally absent.","properties":{"conditionalFormatting":{"items":{"additionalProperties":false,"description":"A closed, governed conditional-format binding over compiled result names.","properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"rule":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"high":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["gradient"],"type":"string"},"low":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"maximum":{"type":"number"},"minimum":{"type":"number"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"}},"required":["high","kind","low","maximum","minimum","nullStyle"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["rules"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"rules":{"items":{"additionalProperties":false,"description":"One ordered numeric comparison in a rules-based conditional format.","properties":{"operator":{"enum":["less_than","less_or_equal","greater_than","greater_or_equal","equal","not_equal"],"type":"string"},"style":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"value":{"type":"number"}},"required":["operator","style","value"],"type":"object"},"type":"array"}},"required":["defaultStyle","kind","nullStyle","rules"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["field"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"values":{"additionalProperties":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"type":"object"}},"required":["defaultStyle","kind","nullStyle","source","values"],"type":"object"}]},"target":{"enum":["mark_fill","mark_stroke","series_color","label_foreground","visual_background","cell_foreground","cell_background","kpi_value","icon"],"type":"string"}},"required":["field","id","rule","target"],"type":"object"},"type":"array"},"type":{"type":"string"}},"required":["type"],"type":"object"}],"description":"Strict point/scatter presentation and its governed result-frame channels.","properties":{"brush":{"items":{"enum":["rectangle","lasso"],"type":"string"},"type":"array"},"color":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"colorScale":{"additionalProperties":false,"description":"A governed result-frame binding used by a point/scatter visual.","properties":{"kind":{"enum":["categorical","quantitative"],"type":"string"},"maximum":{"type":"number"},"minimum":{"type":"number"},"scheme":{"type":"string"}},"required":["kind"],"type":"object"},"identity":{"items":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"type":"array"},"label":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"labels":{"additionalProperties":false,"properties":{"density":{"enum":["hidden","automatic","dense","always"],"type":"string"},"maxCharacters":{"type":"integer"},"minimumSpacing":{"type":"integer"},"priority":{"items":{"enum":["selected","anomaly","threshold"],"type":"string"},"type":"array"},"tooltipFallback":{"type":"boolean"}},"required":["density"],"type":"object"},"legend":{"enum":["none","top","right","bottom","left"],"type":"string"},"overplot":{"additionalProperties":false,"description":"Overplotting policy for a point/scatter visual.","properties":{"largeMode":{"enum":["automatic","always","never"],"type":"string"},"largeThreshold":{"type":"integer"},"opacity":{"type":"number"},"strategy":{"enum":["show_all","opacity"],"type":"string"}},"required":["strategy"],"type":"object"},"series":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"size":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"sizeScale":{"additionalProperties":false,"description":"Pixel and optional domain bounds for a governed point size binding.","properties":{"maximum":{"type":"number"},"maximumPixels":{"type":"number"},"minimum":{"type":"number"},"minimumPixels":{"type":"number"}},"required":["maximumPixels","minimumPixels"],"type":"object"},"tooltip":{"items":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"type":"array"},"type":{"enum":["point"],"type":"string"},"x":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"y":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["identity","type","x","y"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"Renderer-neutral presentation policies; renderer options are intentionally absent.","properties":{"conditionalFormatting":{"items":{"additionalProperties":false,"description":"A closed, governed conditional-format binding over compiled result names.","properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"rule":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"high":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["gradient"],"type":"string"},"low":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"maximum":{"type":"number"},"minimum":{"type":"number"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"}},"required":["high","kind","low","maximum","minimum","nullStyle"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["rules"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"rules":{"items":{"additionalProperties":false,"description":"One ordered numeric comparison in a rules-based conditional format.","properties":{"operator":{"enum":["less_than","less_or_equal","greater_than","greater_or_equal","equal","not_equal"],"type":"string"},"style":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"value":{"type":"number"}},"required":["operator","style","value"],"type":"object"},"type":"array"}},"required":["defaultStyle","kind","nullStyle","rules"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["field"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"values":{"additionalProperties":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"type":"object"}},"required":["defaultStyle","kind","nullStyle","source","values"],"type":"object"}]},"target":{"enum":["mark_fill","mark_stroke","series_color","label_foreground","visual_background","cell_foreground","cell_background","kpi_value","icon"],"type":"string"}},"required":["field","id","rule","target"],"type":"object"},"type":"array"},"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"align":{"enum":["left","center","right"],"type":"string"},"centerLabel":{"type":"string"},"displayUnits":{"enum":["auto","none","thousands","millions","billions","trillions"],"type":"string"},"innerRadius":{"type":"number"},"labelPosition":{"enum":["automatic","inside","outside","top"],"type":"string"},"labels":{"additionalProperties":false,"properties":{"density":{"enum":["hidden","automatic","dense","always"],"type":"string"},"maxCharacters":{"type":"integer"},"minimumSpacing":{"type":"integer"},"priority":{"items":{"enum":["selected","anomaly","threshold"],"type":"string"},"type":"array"},"tooltipFallback":{"type":"boolean"}},"required":["density"],"type":"object"},"legend":{"enum":["none","top","right","bottom","left"],"type":"string"},"orientation":{"enum":["horizontal","vertical"],"type":"string"},"outerRadius":{"type":"number"},"rose":{"type":"boolean"},"sort":{"enum":["ascending","descending"],"type":"string"},"type":{"enum":["proportional"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"Renderer-neutral presentation policies; renderer options are intentionally absent.","properties":{"conditionalFormatting":{"items":{"additionalProperties":false,"description":"A closed, governed conditional-format binding over compiled result names.","properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"rule":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"high":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["gradient"],"type":"string"},"low":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"maximum":{"type":"number"},"minimum":{"type":"number"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"}},"required":["high","kind","low","maximum","minimum","nullStyle"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["rules"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"rules":{"items":{"additionalProperties":false,"description":"One ordered numeric comparison in a rules-based conditional format.","properties":{"operator":{"enum":["less_than","less_or_equal","greater_than","greater_or_equal","equal","not_equal"],"type":"string"},"style":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"value":{"type":"number"}},"required":["operator","style","value"],"type":"object"},"type":"array"}},"required":["defaultStyle","kind","nullStyle","rules"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["field"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"values":{"additionalProperties":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"type":"object"}},"required":["defaultStyle","kind","nullStyle","source","values"],"type":"object"}]},"target":{"enum":["mark_fill","mark_stroke","series_color","label_foreground","visual_background","cell_foreground","cell_background","kpi_value","icon"],"type":"string"}},"required":["field","id","rule","target"],"type":"object"},"type":"array"},"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"breadcrumb":{"type":"boolean"},"curveness":{"type":"number"},"focus":{"enum":["none","adjacency"],"type":"string"},"initialDepth":{"type":"integer"},"labels":{"additionalProperties":false,"properties":{"density":{"enum":["hidden","automatic","dense","always"],"type":"string"},"maxCharacters":{"type":"integer"},"minimumSpacing":{"type":"integer"},"priority":{"items":{"enum":["selected","anomaly","threshold"],"type":"string"},"type":"array"},"tooltipFallback":{"type":"boolean"}},"required":["density"],"type":"object"},"layout":{"enum":["standard","circular"],"type":"string"},"legend":{"enum":["none","top","right","bottom","left"],"type":"string"},"nodeGap":{"type":"number"},"orientation":{"enum":["horizontal","vertical"],"type":"string"},"roam":{"type":"boolean"},"type":{"enum":["hierarchy"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"Renderer-neutral presentation policies; renderer options are intentionally absent.","properties":{"conditionalFormatting":{"items":{"additionalProperties":false,"description":"A closed, governed conditional-format binding over compiled result names.","properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"rule":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"high":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["gradient"],"type":"string"},"low":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"maximum":{"type":"number"},"minimum":{"type":"number"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"}},"required":["high","kind","low","maximum","minimum","nullStyle"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["rules"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"rules":{"items":{"additionalProperties":false,"description":"One ordered numeric comparison in a rules-based conditional format.","properties":{"operator":{"enum":["less_than","less_or_equal","greater_than","greater_or_equal","equal","not_equal"],"type":"string"},"style":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"value":{"type":"number"}},"required":["operator","style","value"],"type":"object"},"type":"array"}},"required":["defaultStyle","kind","nullStyle","rules"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["field"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"values":{"additionalProperties":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"type":"object"}},"required":["defaultStyle","kind","nullStyle","source","values"],"type":"object"}]},"target":{"enum":["mark_fill","mark_stroke","series_color","label_foreground","visual_background","cell_foreground","cell_background","kpi_value","icon"],"type":"string"}},"required":["field","id","rule","target"],"type":"object"},"type":"array"},"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"area":{"type":"boolean"},"displayUnits":{"enum":["auto","none","thousands","millions","billions","trillions"],"type":"string"},"labels":{"additionalProperties":false,"properties":{"density":{"enum":["hidden","automatic","dense","always"],"type":"string"},"maxCharacters":{"type":"integer"},"minimumSpacing":{"type":"integer"},"priority":{"items":{"enum":["selected","anomaly","threshold"],"type":"string"},"type":"array"},"tooltipFallback":{"type":"boolean"}},"required":["density"],"type":"object"},"legend":{"enum":["none","top","right","bottom","left"],"type":"string"},"maximum":{"type":"number"},"minimum":{"type":"number"},"progressWidth":{"type":"number"},"showPointer":{"type":"boolean"},"target":{"type":"number"},"thresholds":{"items":{"additionalProperties":false,"properties":{"tone":{"enum":["neutral","ink","success","warning","danger"],"type":"string"},"value":{"type":"number"}},"required":["tone","value"],"type":"object"},"type":"array"},"type":{"enum":["polar"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"Renderer-neutral presentation policies; renderer options are intentionally absent.","properties":{"conditionalFormatting":{"items":{"additionalProperties":false,"description":"A closed, governed conditional-format binding over compiled result names.","properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"rule":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"high":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["gradient"],"type":"string"},"low":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"maximum":{"type":"number"},"minimum":{"type":"number"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"}},"required":["high","kind","low","maximum","minimum","nullStyle"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["rules"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"rules":{"items":{"additionalProperties":false,"description":"One ordered numeric comparison in a rules-based conditional format.","properties":{"operator":{"enum":["less_than","less_or_equal","greater_than","greater_or_equal","equal","not_equal"],"type":"string"},"style":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"value":{"type":"number"}},"required":["operator","style","value"],"type":"object"},"type":"array"}},"required":["defaultStyle","kind","nullStyle","rules"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["field"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"values":{"additionalProperties":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"type":"object"}},"required":["defaultStyle","kind","nullStyle","source","values"],"type":"object"}]},"target":{"enum":["mark_fill","mark_stroke","series_color","label_foreground","visual_background","cell_foreground","cell_background","kpi_value","icon"],"type":"string"}},"required":["field","id","rule","target"],"type":"object"},"type":"array"},"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"basemap":{"type":"string"},"camera":{"additionalProperties":false,"properties":{"center":{"items":{"type":"number"},"type":"array"},"maximumZoom":{"type":"number"},"minimumZoom":{"type":"number"},"mode":{"enum":["fit_data","fixed","preserve"],"type":"string"},"padding":{"type":"integer"},"zoom":{"type":"number"}},"type":"object"},"controls":{"additionalProperties":false,"properties":{"compass":{"type":"boolean"},"reset":{"type":"boolean"},"zoom":{"type":"boolean"}},"type":"object"},"labelDensity":{"enum":["hidden","normal","dense"],"type":"string"},"labels":{"additionalProperties":false,"properties":{"density":{"enum":["hidden","automatic","dense","always"],"type":"string"},"maxCharacters":{"type":"integer"},"minimumSpacing":{"type":"integer"},"priority":{"items":{"enum":["selected","anomaly","threshold"],"type":"string"},"type":"array"},"tooltipFallback":{"type":"boolean"}},"required":["density"],"type":"object"},"layers":{"items":{"oneOf":[{"additionalProperties":false,"allOf":[{"allOf":[{"properties":{"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"maximumZoom":{"type":"number"},"minimumZoom":{"type":"number"},"position":{"enum":["below_labels","above_labels"],"type":"string"},"tooltip":{"items":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"type":"array"}},"required":["id"],"type":"object"}],"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"category":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"cluster":{"additionalProperties":false,"properties":{"enabled":{"type":"boolean"},"maximumZoom":{"type":"integer"},"minimumPoints":{"type":"integer"},"radius":{"type":"integer"},"showCount":{"type":"boolean"}},"type":"object"},"color":{"additionalProperties":false,"properties":{"domainMaximum":{"type":"number"},"domainMidpoint":{"type":"number"},"domainMinimum":{"type":"number"},"kind":{"enum":["sequential","diverging","categorical"],"type":"string"},"nullColor":{"type":"string"},"palette":{"type":"string"},"reverse":{"type":"boolean"}},"type":"object"},"kind":{"enum":["point"],"type":"string"},"latitude":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"longitude":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"opacity":{"type":"number"},"size":{"additionalProperties":false,"properties":{"domainMaximum":{"type":"number"},"domainMinimum":{"type":"number"},"maximumRadius":{"type":"number"},"minimumRadius":{"type":"number"}},"type":"object"},"stroke":{"additionalProperties":false,"properties":{"color":{"type":"string"},"opacity":{"type":"number"},"width":{"type":"number"}},"type":"object"},"value":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["kind","latitude","longitude"],"type":"object"},{"additionalProperties":false,"allOf":[{"allOf":[{"properties":{"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"maximumZoom":{"type":"number"},"minimumZoom":{"type":"number"},"position":{"enum":["below_labels","above_labels"],"type":"string"},"tooltip":{"items":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"type":"array"}},"required":["id"],"type":"object"}],"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"category":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"color":{"additionalProperties":false,"properties":{"domainMaximum":{"type":"number"},"domainMidpoint":{"type":"number"},"domainMinimum":{"type":"number"},"kind":{"enum":["sequential","diverging","categorical"],"type":"string"},"nullColor":{"type":"string"},"palette":{"type":"string"},"reverse":{"type":"boolean"}},"type":"object"},"geometryAsset":{"type":"string"},"join":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"kind":{"enum":["choropleth"],"type":"string"},"opacity":{"type":"number"},"stroke":{"additionalProperties":false,"properties":{"color":{"type":"string"},"opacity":{"type":"number"},"width":{"type":"number"}},"type":"object"},"value":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["geometryAsset","join","kind"],"type":"object"},{"additionalProperties":false,"allOf":[{"allOf":[{"properties":{"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"maximumZoom":{"type":"number"},"minimumZoom":{"type":"number"},"position":{"enum":["below_labels","above_labels"],"type":"string"},"tooltip":{"items":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"type":"array"}},"required":["id"],"type":"object"}],"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"color":{"additionalProperties":false,"properties":{"domainMaximum":{"type":"number"},"domainMidpoint":{"type":"number"},"domainMinimum":{"type":"number"},"kind":{"enum":["sequential","diverging","categorical"],"type":"string"},"nullColor":{"type":"string"},"palette":{"type":"string"},"reverse":{"type":"boolean"}},"type":"object"},"geometryAsset":{"type":"string"},"kind":{"enum":["reference"],"type":"string"},"opacity":{"type":"number"},"stroke":{"additionalProperties":false,"properties":{"color":{"type":"string"},"opacity":{"type":"number"},"width":{"type":"number"}},"type":"object"}},"required":["geometryAsset","kind"],"type":"object"},{"additionalProperties":false,"allOf":[{"allOf":[{"properties":{"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"maximumZoom":{"type":"number"},"minimumZoom":{"type":"number"},"position":{"enum":["below_labels","above_labels"],"type":"string"},"tooltip":{"items":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"type":"array"}},"required":["id"],"type":"object"}],"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"color":{"additionalProperties":false,"properties":{"domainMaximum":{"type":"number"},"domainMidpoint":{"type":"number"},"domainMinimum":{"type":"number"},"kind":{"enum":["sequential","diverging","categorical"],"type":"string"},"nullColor":{"type":"string"},"palette":{"type":"string"},"reverse":{"type":"boolean"}},"type":"object"},"heat":{"additionalProperties":false,"properties":{"intensity":{"type":"number"},"radius":{"type":"number"}},"type":"object"},"kind":{"enum":["heat"],"type":"string"},"latitude":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"longitude":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"opacity":{"type":"number"},"value":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["kind","latitude","longitude"],"type":"object"},{"additionalProperties":false,"allOf":[{"allOf":[{"properties":{"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"maximumZoom":{"type":"number"},"minimumZoom":{"type":"number"},"position":{"enum":["below_labels","above_labels"],"type":"string"},"tooltip":{"items":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"type":"array"}},"required":["id"],"type":"object"}],"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"color":{"additionalProperties":false,"properties":{"domainMaximum":{"type":"number"},"domainMidpoint":{"type":"number"},"domainMinimum":{"type":"number"},"kind":{"enum":["sequential","diverging","categorical"],"type":"string"},"nullColor":{"type":"string"},"palette":{"type":"string"},"reverse":{"type":"boolean"}},"type":"object"},"heat":{"additionalProperties":false,"properties":{"intensity":{"type":"number"},"radius":{"type":"number"}},"type":"object"},"kind":{"enum":["density"],"type":"string"},"latitude":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"longitude":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"opacity":{"type":"number"},"value":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["kind","latitude","longitude"],"type":"object"},{"additionalProperties":false,"allOf":[{"allOf":[{"properties":{"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"maximumZoom":{"type":"number"},"minimumZoom":{"type":"number"},"position":{"enum":["below_labels","above_labels"],"type":"string"},"tooltip":{"items":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"type":"array"}},"required":["id"],"type":"object"}],"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"category":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"color":{"additionalProperties":false,"properties":{"domainMaximum":{"type":"number"},"domainMidpoint":{"type":"number"},"domainMinimum":{"type":"number"},"kind":{"enum":["sequential","diverging","categorical"],"type":"string"},"nullColor":{"type":"string"},"palette":{"type":"string"},"reverse":{"type":"boolean"}},"type":"object"},"kind":{"enum":["path"],"type":"string"},"latitude":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"line":{"additionalProperties":false,"properties":{"curvature":{"type":"number"},"width":{"type":"number"}},"type":"object"},"longitude":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"opacity":{"type":"number"},"order":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"path":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"stroke":{"additionalProperties":false,"properties":{"color":{"type":"string"},"opacity":{"type":"number"},"width":{"type":"number"}},"type":"object"},"value":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["kind","latitude","longitude","order","path"],"type":"object"}]},"type":"array"},"roam":{"type":"boolean"},"theme":{"enum":["auto","light","dark"],"type":"string"},"type":{"enum":["geographic"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"Renderer-neutral presentation policies; renderer options are intentionally absent.","properties":{"conditionalFormatting":{"items":{"additionalProperties":false,"description":"A closed, governed conditional-format binding over compiled result names.","properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"rule":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"high":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["gradient"],"type":"string"},"low":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"maximum":{"type":"number"},"minimum":{"type":"number"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"}},"required":["high","kind","low","maximum","minimum","nullStyle"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["rules"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"rules":{"items":{"additionalProperties":false,"description":"One ordered numeric comparison in a rules-based conditional format.","properties":{"operator":{"enum":["less_than","less_or_equal","greater_than","greater_or_equal","equal","not_equal"],"type":"string"},"style":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"value":{"type":"number"}},"required":["operator","style","value"],"type":"object"},"type":"array"}},"required":["defaultStyle","kind","nullStyle","rules"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["field"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"values":{"additionalProperties":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"type":"object"}},"required":["defaultStyle","kind","nullStyle","source","values"],"type":"object"}]},"target":{"enum":["mark_fill","mark_stroke","series_color","label_foreground","visual_background","cell_foreground","cell_background","kpi_value","icon"],"type":"string"}},"required":["field","id","rule","target"],"type":"object"},"type":"array"},"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"rowHeight":{"type":"integer"},"showHeader":{"type":"boolean"},"striped":{"type":"boolean"},"type":{"enum":["table"],"type":"string"}},"required":["rowHeight","showHeader","striped","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"Renderer-neutral presentation policies; renderer options are intentionally absent.","properties":{"conditionalFormatting":{"items":{"additionalProperties":false,"description":"A closed, governed conditional-format binding over compiled result names.","properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"rule":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"high":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["gradient"],"type":"string"},"low":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"maximum":{"type":"number"},"minimum":{"type":"number"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"}},"required":["high","kind","low","maximum","minimum","nullStyle"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["rules"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"rules":{"items":{"additionalProperties":false,"description":"One ordered numeric comparison in a rules-based conditional format.","properties":{"operator":{"enum":["less_than","less_or_equal","greater_than","greater_or_equal","equal","not_equal"],"type":"string"},"style":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"value":{"type":"number"}},"required":["operator","style","value"],"type":"object"},"type":"array"}},"required":["defaultStyle","kind","nullStyle","rules"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["field"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"values":{"additionalProperties":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"type":"object"}},"required":["defaultStyle","kind","nullStyle","source","values"],"type":"object"}]},"target":{"enum":["mark_fill","mark_stroke","series_color","label_foreground","visual_background","cell_foreground","cell_background","kpi_value","icon"],"type":"string"}},"required":["field","id","rule","target"],"type":"object"},"type":"array"},"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"comparison":{"additionalProperties":false,"description":"A governed numeric result used for a KPI comparison or goal.","properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"label":{"type":"string"},"reducer":{"enum":["first","last","minimum","maximum","mean","median"],"type":"string"}},"required":["dataset","field","label"],"type":"object"},"delta":{"enum":["absolute","relative"],"type":"string"},"displayUnits":{"enum":["auto","none","thousands","millions","billions","trillions"],"type":"string"},"favorableDirection":{"enum":["increase","decrease","neutral"],"type":"string"},"goal":{"additionalProperties":false,"description":"A governed numeric result used for a KPI comparison or goal.","properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"label":{"type":"string"},"reducer":{"enum":["first","last","minimum","maximum","mean","median"],"type":"string"}},"required":["dataset","field","label"],"type":"object"},"missingComparison":{"enum":["show_unavailable","hide"],"type":"string"},"mode":{"enum":["compact","bullet","progress"],"type":"string"},"note":{"type":"string"},"ranges":{"items":{"additionalProperties":false,"properties":{"label":{"type":"string"},"maximum":{"type":"number"},"minimum":{"type":"number"},"tone":{"enum":["neutral","ink","success","warning","danger"],"type":"string"}},"required":["label","tone"],"type":"object"},"type":"array"},"thresholds":{"items":{"additionalProperties":false,"properties":{"tone":{"enum":["neutral","ink","success","warning","danger"],"type":"string"},"value":{"type":"number"}},"required":["tone","value"],"type":"object"},"type":"array"},"tone":{"enum":["neutral","ink","success","warning","danger"],"type":"string"},"trend":{"additionalProperties":false,"description":"A governed result frame used for a KPI sparkline trend.","properties":{"category":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"value":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["category","dataset","value"],"type":"object"},"type":{"enum":["kpi"],"type":"string"}},"required":["type"],"type":"object"}]},"query":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"dimensions":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic dimension reference. Physical dataset fields are not accepted.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"dimension":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"grain":{"enum":["second","minute","hour","day","week","month","quarter","year"],"type":"string"}},"required":["dimension"],"type":"object"}]},"type":"array"},"limit":{"type":"integer"},"metrics":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic metric reference. Metric expressions are never part of this type.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"metric":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["metric"],"type":"object"}]},"type":"array"},"sort":{"items":{"additionalProperties":false,"properties":{"direction":{"enum":["asc","desc"],"type":"string"},"field":{"description":"A compiled result-frame field name, never a source member or expression.","minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["direction","field"],"type":"object"},"type":"array"},"type":{"enum":["aggregate"],"type":"string"}},"required":["dimensions","metrics","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"fields":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A physical field reference is only valid for a records query root dataset.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["field"],"type":"object"}]},"type":"array"},"limit":{"type":"integer"},"sort":{"items":{"additionalProperties":false,"properties":{"direction":{"enum":["asc","desc"],"type":"string"},"field":{"description":"A compiled result-frame field name, never a source member or expression.","minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["direction","field"],"type":"object"},"type":"array"},"type":{"enum":["records"],"type":"string"}},"required":["dataset","fields","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"columns":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic dimension reference. Physical dataset fields are not accepted.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"dimension":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"grain":{"enum":["second","minute","hour","day","week","month","quarter","year"],"type":"string"}},"required":["dimension"],"type":"object"}]},"type":"array"},"metrics":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic metric reference. Metric expressions are never part of this type.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"metric":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["metric"],"type":"object"}]},"type":"array"},"rows":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic dimension reference. Physical dataset fields are not accepted.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"dimension":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"grain":{"enum":["second","minute","hour","day","week","month","quarter","year"],"type":"string"}},"required":["dimension"],"type":"object"}]},"type":"array"},"sort":{"items":{"additionalProperties":false,"properties":{"direction":{"enum":["asc","desc"],"type":"string"},"field":{"description":"A compiled result-frame field name, never a source member or expression.","minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["direction","field"],"type":"object"},"type":"array"},"totals":{"additionalProperties":false,"properties":{"columns":{"type":"boolean"},"grand":{"type":"boolean"},"rows":{"type":"boolean"}},"type":"object"},"type":{"enum":["pivot"],"type":"string"},"window":{"additionalProperties":false,"properties":{"limit":{"type":"integer"},"offset":{"type":"integer"}},"required":["limit"],"type":"object"}},"required":["columns","metrics","rows","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"approximation":{"enum":["exact","approximate"],"type":"string"},"bins":{"type":"integer"},"domain":{"additionalProperties":false,"properties":{"maximum":{"type":"number"},"minimum":{"type":"number"}},"type":"object"},"field":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic metric reference. Metric expressions are never part of this type.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"metric":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["metric"],"type":"object"}]},"nullPolicy":{"enum":["omit","include"],"type":"string"},"type":{"enum":["histogram"],"type":"string"}},"required":["approximation","bins","field","nullPolicy","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"approximation":{"enum":["exact","approximate"],"type":"string"},"field":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic metric reference. Metric expressions are never part of this type.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"metric":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["metric"],"type":"object"}]},"group":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic dimension reference. Physical dataset fields are not accepted.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"dimension":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"grain":{"enum":["second","minute","hour","day","week","month","quarter","year"],"type":"string"}},"required":["dimension"],"type":"object"}]},"limit":{"type":"integer"},"outliers":{"enum":["omit","include"],"type":"string"},"quantiles":{"items":{"type":"number"},"type":"array"},"type":{"enum":["distribution"],"type":"string"},"whiskers":{"additionalProperties":false,"properties":{"lower":{"type":"number"},"upper":{"type":"number"}},"required":["lower","upper"],"type":"object"}},"required":["approximation","field","outliers","quantiles","type"],"type":"object"}]},"subtitle":{"type":"string"},"title":{"type":"string"},"type":{"enum":["line","area","bar","column","pie","donut","scatter","funnel","treemap","gauge","heatmap","sankey","graph","map","candlestick","boxplot","combo","waterfall","histogram","radar","tree","sunburst","kpi","table","matrix","pivot"],"type":"string"}},"required":["presentation","query","type"],"type":"object"}},"required":["semanticModelId","visual"],"type":"object"}`
+	QueryVisualInputSchemaJSON = `{"additionalProperties":false,"properties":{"filters":{"description":"Governed semantic filters applied before visualization. Inline data and arbitrary expressions are not accepted.","items":{"additionalProperties":false,"properties":{"control":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"options":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"dependsOn":{"items":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"type":"array"},"includeNull":{"type":"boolean"},"limit":{"type":"integer"},"type":{"enum":["distinct"],"type":"string"}},"required":["dataset","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["static"],"type":"string"},"values":{"items":{"additionalProperties":false,"properties":{"label":{"type":"string"},"value":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["string"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["boolean"],"type":"string"},"value":{"type":"boolean"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["integer"],"type":"string"},"value":{"description":"Precision-safe integer representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["decimal"],"type":"string"},"value":{"description":"Canonical decimal representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["date"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["timestamp"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"}]}},"required":["label","value"],"type":"object"},"type":"array"}},"required":["type","values"],"type":"object"}]},"type":{"enum":["singleSelect"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"maxSelectedValues":{"type":"integer"},"options":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"dependsOn":{"items":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"type":"array"},"includeNull":{"type":"boolean"},"limit":{"type":"integer"},"type":{"enum":["distinct"],"type":"string"}},"required":["dataset","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["static"],"type":"string"},"values":{"items":{"additionalProperties":false,"properties":{"label":{"type":"string"},"value":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["string"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["boolean"],"type":"string"},"value":{"type":"boolean"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["integer"],"type":"string"},"value":{"description":"Precision-safe integer representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["decimal"],"type":"string"},"value":{"description":"Canonical decimal representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["date"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["timestamp"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"}]}},"required":["label","value"],"type":"object"},"type":"array"}},"required":["type","values"],"type":"object"}]},"type":{"enum":["multiSelect"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["text"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["numericRange"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["dateRange"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["relativePeriod"],"type":"string"}},"required":["type"],"type":"object"}]},"default":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["unfiltered"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"operator":{"enum":["isNull","isNotNull","in","notIn","equals","notEquals","contains","notContains","startsWith","endsWith","greaterThan","greaterThanOrEqual","lessThan","lessThanOrEqual"],"type":"string"},"type":{"enum":["nullCheck"],"type":"string"}},"required":["operator","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"operator":{"enum":["isNull","isNotNull","in","notIn","equals","notEquals","contains","notContains","startsWith","endsWith","greaterThan","greaterThanOrEqual","lessThan","lessThanOrEqual"],"type":"string"},"type":{"enum":["set"],"type":"string"},"values":{"items":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["string"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["boolean"],"type":"string"},"value":{"type":"boolean"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["integer"],"type":"string"},"value":{"description":"Precision-safe integer representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["decimal"],"type":"string"},"value":{"description":"Canonical decimal representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["date"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["timestamp"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"}]},"type":"array"}},"required":["operator","type","values"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"operator":{"enum":["isNull","isNotNull","in","notIn","equals","notEquals","contains","notContains","startsWith","endsWith","greaterThan","greaterThanOrEqual","lessThan","lessThanOrEqual"],"type":"string"},"type":{"enum":["comparison"],"type":"string"},"value":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["string"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["boolean"],"type":"string"},"value":{"type":"boolean"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["integer"],"type":"string"},"value":{"description":"Precision-safe integer representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["decimal"],"type":"string"},"value":{"description":"Canonical decimal representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["date"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["timestamp"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"}]}},"required":["operator","type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"lower":{"additionalProperties":false,"properties":{"inclusive":{"type":"boolean"},"value":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["string"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["boolean"],"type":"string"},"value":{"type":"boolean"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["integer"],"type":"string"},"value":{"description":"Precision-safe integer representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["decimal"],"type":"string"},"value":{"description":"Canonical decimal representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["date"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["timestamp"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"}]}},"required":["inclusive","value"],"type":"object"},"type":{"enum":["range"],"type":"string"},"upper":{"additionalProperties":false,"properties":{"inclusive":{"type":"boolean"},"value":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["string"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["boolean"],"type":"string"},"value":{"type":"boolean"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["integer"],"type":"string"},"value":{"description":"Precision-safe integer representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["decimal"],"type":"string"},"value":{"description":"Canonical decimal representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["date"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["timestamp"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"}]}},"required":["inclusive","value"],"type":"object"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"anchor":{"enum":["currentTime","firstAvailable","lastAvailable","fixed"],"type":"string"},"anchorValue":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["string"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["boolean"],"type":"string"},"value":{"type":"boolean"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["integer"],"type":"string"},"value":{"description":"Precision-safe integer representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["decimal"],"type":"string"},"value":{"description":"Canonical decimal representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["date"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["timestamp"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"}]},"count":{"type":"integer"},"direction":{"enum":["previous","current","next"],"type":"string"},"includeCurrent":{"type":"boolean"},"type":{"enum":["relativePeriod"],"type":"string"},"unit":{"enum":["minute","hour","day","week","month","quarter","year"],"type":"string"}},"required":["anchor","count","direction","includeCurrent","type","unit"],"type":"object"}]},"description":{"type":"string"},"dimension":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"type":"string"},"operators":{"items":{"enum":["isNull","isNotNull","in","notIn","equals","notEquals","contains","notContains","startsWith","endsWith","greaterThan","greaterThanOrEqual","lessThan","lessThanOrEqual"],"type":"string"},"type":"array"},"readerEditable":{"type":"boolean"},"required":{"type":"boolean"},"targets":{"items":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"type":"array"},"urlParameter":{"maxLength":64,"minLength":1,"pattern":"^[A-Za-z][A-Za-z0-9_-]{0,63}$","type":"string"}},"required":["control","dimension","id","label"],"type":"object"},"type":"array"},"semanticModelId":{"description":"Stable semantic_model resource ID returned by catalog_search/catalog_get.","minLength":1,"type":"string"},"visual":{"additionalProperties":false,"properties":{"accessibility":{"additionalProperties":false,"properties":{"announceChanges":{"type":"boolean"},"description":{"type":"string"},"summary":{"type":"string"},"title":{"type":"string"}},"type":"object"},"calculations":{"items":{"additionalProperties":false,"properties":{"axis":{"enum":["rows","columns","hierarchy","facets"],"type":"string"},"hidden":{"type":"boolean"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"type":"string"},"lookup":{"additionalProperties":false,"properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"value":{"type":"string"}},"required":["field","value"],"type":"object"},"offset":{"type":"integer"},"orderBy":{"items":{"additionalProperties":false,"properties":{"direction":{"enum":["asc","desc"],"type":"string"},"field":{"description":"A compiled result-frame field name, never a source member or expression.","minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["direction","field"],"type":"object"},"type":"array"},"parent":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"partitionBy":{"items":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"type":"array"},"reset":{"enum":["none","highest_parent","lowest_parent"],"type":"string"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"template":{"enum":["running_total","moving_average","difference","percentage_difference","percent_of_parent","percent_of_grand_total","rank","cumulative_contribution","lookup"],"type":"string"},"window":{"type":"integer"}},"required":["id","source","template"],"type":"object"},"type":"array"},"dataBudget":{"additionalProperties":false,"properties":{"maxRows":{"type":"integer"},"requiredCompleteness":{"enum":["complete","truncated","partial","empty"],"type":"string"}},"required":["maxRows"],"type":"object"},"datasets":{"additionalProperties":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"dimensions":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic dimension reference. Physical dataset fields are not accepted.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"dimension":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"grain":{"enum":["second","minute","hour","day","week","month","quarter","year"],"type":"string"}},"required":["dimension"],"type":"object"}]},"type":"array"},"limit":{"type":"integer"},"metrics":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic metric reference. Metric expressions are never part of this type.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"metric":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["metric"],"type":"object"}]},"type":"array"},"sort":{"items":{"additionalProperties":false,"properties":{"direction":{"enum":["asc","desc"],"type":"string"},"field":{"description":"A compiled result-frame field name, never a source member or expression.","minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["direction","field"],"type":"object"},"type":"array"},"type":{"enum":["aggregate"],"type":"string"}},"required":["dimensions","metrics","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"fields":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A physical field reference is only valid for a records query root dataset.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["field"],"type":"object"}]},"type":"array"},"limit":{"type":"integer"},"sort":{"items":{"additionalProperties":false,"properties":{"direction":{"enum":["asc","desc"],"type":"string"},"field":{"description":"A compiled result-frame field name, never a source member or expression.","minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["direction","field"],"type":"object"},"type":"array"},"type":{"enum":["records"],"type":"string"}},"required":["dataset","fields","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"columns":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic dimension reference. Physical dataset fields are not accepted.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"dimension":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"grain":{"enum":["second","minute","hour","day","week","month","quarter","year"],"type":"string"}},"required":["dimension"],"type":"object"}]},"type":"array"},"metrics":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic metric reference. Metric expressions are never part of this type.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"metric":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["metric"],"type":"object"}]},"type":"array"},"rows":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic dimension reference. Physical dataset fields are not accepted.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"dimension":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"grain":{"enum":["second","minute","hour","day","week","month","quarter","year"],"type":"string"}},"required":["dimension"],"type":"object"}]},"type":"array"},"sort":{"items":{"additionalProperties":false,"properties":{"direction":{"enum":["asc","desc"],"type":"string"},"field":{"description":"A compiled result-frame field name, never a source member or expression.","minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["direction","field"],"type":"object"},"type":"array"},"totals":{"additionalProperties":false,"properties":{"columns":{"type":"boolean"},"grand":{"type":"boolean"},"rows":{"type":"boolean"}},"type":"object"},"type":{"enum":["pivot"],"type":"string"},"window":{"additionalProperties":false,"properties":{"limit":{"type":"integer"},"offset":{"type":"integer"}},"required":["limit"],"type":"object"}},"required":["columns","metrics","rows","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"approximation":{"enum":["exact","approximate"],"type":"string"},"bins":{"type":"integer"},"domain":{"additionalProperties":false,"properties":{"maximum":{"type":"number"},"minimum":{"type":"number"}},"type":"object"},"field":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic metric reference. Metric expressions are never part of this type.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"metric":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["metric"],"type":"object"}]},"nullPolicy":{"enum":["omit","include"],"type":"string"},"type":{"enum":["histogram"],"type":"string"}},"required":["approximation","bins","field","nullPolicy","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"approximation":{"enum":["exact","approximate"],"type":"string"},"field":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic metric reference. Metric expressions are never part of this type.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"metric":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["metric"],"type":"object"}]},"group":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic dimension reference. Physical dataset fields are not accepted.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"dimension":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"grain":{"enum":["second","minute","hour","day","week","month","quarter","year"],"type":"string"}},"required":["dimension"],"type":"object"}]},"limit":{"type":"integer"},"outliers":{"enum":["omit","include"],"type":"string"},"quantiles":{"items":{"type":"number"},"type":"array"},"type":{"enum":["distribution"],"type":"string"},"whiskers":{"additionalProperties":false,"properties":{"lower":{"type":"number"},"upper":{"type":"number"}},"required":["lower","upper"],"type":"object"}},"required":["approximation","field","outliers","quantiles","type"],"type":"object"}]},"propertyNames":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"type":"object"},"description":{"type":"string"},"interactions":{"items":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"targets":{"items":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"type":"array"},"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"highlightTargets":{"items":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"type":"array"},"mappings":{"items":{"additionalProperties":false,"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"grain":{"enum":["second","minute","hour","day","week","month","quarter","year"],"type":"string"},"label":{"type":"string"},"value":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["field","value"],"type":"object"},"type":"array"},"mode":{"enum":["single","multiple"],"type":"string"},"noneTargets":{"items":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"type":"array"},"toggle":{"type":"boolean"},"type":{"enum":["selection"],"type":"string"}},"required":["mappings","mode","toggle","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"targets":{"items":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"type":"array"},"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"gestures":{"items":{"enum":["box","lasso","radius"],"type":"string"},"type":"array"},"highlightTargets":{"items":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"type":"array"},"latitude":{"additionalProperties":false,"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["field","source"],"type":"object"},"longitude":{"additionalProperties":false,"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["field","source"],"type":"object"},"noneTargets":{"items":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"type":"array"},"type":{"enum":["spatialSelection"],"type":"string"}},"required":["gestures","latitude","longitude","type"],"type":"object"}]},"type":"array"},"metadata":{"additionalProperties":false,"properties":{"description":{"additionalProperties":false,"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"fallback":{"type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"prefix":{"type":"string"},"reducer":{"enum":["first","last","minimum","maximum","mean","median"],"type":"string"},"suffix":{"type":"string"}},"required":["dataset","field"],"type":"object"},"subtitle":{"additionalProperties":false,"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"fallback":{"type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"prefix":{"type":"string"},"reducer":{"enum":["first","last","minimum","maximum","mean","median"],"type":"string"},"suffix":{"type":"string"}},"required":["dataset","field"],"type":"object"},"summary":{"additionalProperties":false,"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"fallback":{"type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"prefix":{"type":"string"},"reducer":{"enum":["first","last","minimum","maximum","mean","median"],"type":"string"},"suffix":{"type":"string"}},"required":["dataset","field"],"type":"object"},"title":{"additionalProperties":false,"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"fallback":{"type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"prefix":{"type":"string"},"reducer":{"enum":["first","last","minimum","maximum","mean","median"],"type":"string"},"suffix":{"type":"string"}},"required":["dataset","field"],"type":"object"}},"type":"object"},"presentation":{"oneOf":[{"additionalProperties":false,"allOf":[{"description":"Renderer-neutral presentation policies; renderer options are intentionally absent.","properties":{"conditionalFormatting":{"items":{"additionalProperties":false,"description":"A closed, governed conditional-format binding over compiled result names.","properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"rule":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"high":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["gradient"],"type":"string"},"low":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"maximum":{"type":"number"},"minimum":{"type":"number"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"}},"required":["high","kind","low","maximum","minimum","nullStyle"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["rules"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"rules":{"items":{"additionalProperties":false,"description":"One ordered numeric comparison in a rules-based conditional format.","properties":{"operator":{"enum":["less_than","less_or_equal","greater_than","greater_or_equal","equal","not_equal"],"type":"string"},"style":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"value":{"type":"number"}},"required":["operator","style","value"],"type":"object"},"type":"array"}},"required":["defaultStyle","kind","nullStyle","rules"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["field"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"values":{"additionalProperties":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"type":"object"}},"required":["defaultStyle","kind","nullStyle","source","values"],"type":"object"}]},"target":{"enum":["mark_fill","mark_stroke","series_color","label_foreground","visual_background","cell_foreground","cell_background","kpi_value","icon"],"type":"string"}},"required":["field","id","rule","target"],"type":"object"},"type":"array"},"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"axes":{"items":{"additionalProperties":false,"description":"A renderer-neutral axis policy for Cartesian and point visuals.","properties":{"displayUnits":{"enum":["auto","none","thousands","millions","billions","trillions"],"type":"string"},"id":{"enum":["x","primary_y","secondary_y"],"type":"string"},"maximum":{"type":"number"},"minimum":{"type":"number"},"scale":{"enum":["automatic","linear","log"],"type":"string"},"tickDensity":{"enum":["automatic","sparse","normal","dense"],"type":"string"},"title":{"type":"string"},"unit":{"type":"string"},"zero":{"enum":["automatic","include","exclude"],"type":"string"}},"required":["id","scale","tickDensity","zero"],"type":"object"},"type":"array"},"dataZoom":{"type":"boolean"},"displayUnits":{"enum":["auto","none","thousands","millions","billions","trillions"],"type":"string"},"eventAnnotations":{"items":{"additionalProperties":false,"description":"A governed event marker, restricted to the visual's horizontal axis.","properties":{"axis":{"enum":["x","primary_y","secondary_y"],"type":"string"},"description":{"type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"type":"string"},"tone":{"enum":["neutral","ink","success","warning","danger"],"type":"string"},"value":{"oneOf":[{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"kind":{"enum":["number"],"type":"string"},"value":{"type":"number"}},"required":["kind","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"kind":{"enum":["text"],"type":"string"},"value":{"type":"string"}},"required":["kind","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"kind":{"enum":["field"],"type":"string"},"reducer":{"enum":["first","last","minimum","maximum","mean","median"],"type":"string"}},"required":["field","kind","reducer"],"type":"object"}]}},"required":["axis","id","label","tone","value"],"type":"object"},"type":"array"},"labelPosition":{"enum":["automatic","inside","outside","top"],"type":"string"},"labels":{"additionalProperties":false,"properties":{"density":{"enum":["hidden","automatic","dense","always"],"type":"string"},"maxCharacters":{"type":"integer"},"minimumSpacing":{"type":"integer"},"priority":{"items":{"enum":["selected","anomaly","threshold"],"type":"string"},"type":"array"},"tooltipFallback":{"type":"boolean"}},"required":["density"],"type":"object"},"legend":{"enum":["none","top","right","bottom","left"],"type":"string"},"orientation":{"enum":["horizontal","vertical"],"type":"string"},"referenceBands":{"items":{"additionalProperties":false,"description":"A governed interval between two horizontal or numeric reference values.","properties":{"axis":{"enum":["x","primary_y","secondary_y"],"type":"string"},"from":{"oneOf":[{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"kind":{"enum":["number"],"type":"string"},"value":{"type":"number"}},"required":["kind","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"kind":{"enum":["text"],"type":"string"},"value":{"type":"string"}},"required":["kind","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"kind":{"enum":["field"],"type":"string"},"reducer":{"enum":["first","last","minimum","maximum","mean","median"],"type":"string"}},"required":["field","kind","reducer"],"type":"object"}]},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"type":"string"},"to":{"oneOf":[{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"kind":{"enum":["number"],"type":"string"},"value":{"type":"number"}},"required":["kind","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"kind":{"enum":["text"],"type":"string"},"value":{"type":"string"}},"required":["kind","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"kind":{"enum":["field"],"type":"string"},"reducer":{"enum":["first","last","minimum","maximum","mean","median"],"type":"string"}},"required":["field","kind","reducer"],"type":"object"}]},"tone":{"enum":["neutral","ink","success","warning","danger"],"type":"string"}},"required":["axis","from","id","to","tone"],"type":"object"},"type":"array"},"referenceLines":{"items":{"additionalProperties":false,"description":"A governed horizontal or numeric reference marker.","properties":{"axis":{"enum":["x","primary_y","secondary_y"],"type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"type":"string"},"tone":{"enum":["neutral","ink","success","warning","danger"],"type":"string"},"value":{"oneOf":[{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"kind":{"enum":["number"],"type":"string"},"value":{"type":"number"}},"required":["kind","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"kind":{"enum":["text"],"type":"string"},"value":{"type":"string"}},"required":["kind","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"kind":{"enum":["field"],"type":"string"},"reducer":{"enum":["first","last","minimum","maximum","mean","median"],"type":"string"}},"required":["field","kind","reducer"],"type":"object"}]}},"required":["axis","id","tone","value"],"type":"object"},"type":"array"},"series":{"description":"Explicit per-measure marks and axis assignments for combo visuals.","items":{"additionalProperties":false,"description":"One result-frame measure and its governed combo rendering policy.","properties":{"axis":{"description":"Axis assignment for one governed combo measure.","enum":["primary","secondary"],"type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"mark":{"description":"Marks supported by a governed multi-measure combo series.","enum":["line","area","bar","column"],"type":"string"}},"required":["axis","field","mark"],"type":"object"},"type":"array"},"showSymbols":{"type":"boolean"},"smooth":{"type":"boolean"},"stacking":{"enum":["none","normal","percent"],"type":"string"},"step":{"description":"Render line segments as discrete steps between categories.","type":"boolean"},"symbolSize":{"type":"number"},"type":{"enum":["cartesian"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"Renderer-neutral presentation policies; renderer options are intentionally absent.","properties":{"conditionalFormatting":{"items":{"additionalProperties":false,"description":"A closed, governed conditional-format binding over compiled result names.","properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"rule":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"high":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["gradient"],"type":"string"},"low":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"maximum":{"type":"number"},"minimum":{"type":"number"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"}},"required":["high","kind","low","maximum","minimum","nullStyle"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["rules"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"rules":{"items":{"additionalProperties":false,"description":"One ordered numeric comparison in a rules-based conditional format.","properties":{"operator":{"enum":["less_than","less_or_equal","greater_than","greater_or_equal","equal","not_equal"],"type":"string"},"style":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"value":{"type":"number"}},"required":["operator","style","value"],"type":"object"},"type":"array"}},"required":["defaultStyle","kind","nullStyle","rules"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["field"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"values":{"additionalProperties":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"type":"object"}},"required":["defaultStyle","kind","nullStyle","source","values"],"type":"object"}]},"target":{"enum":["mark_fill","mark_stroke","series_color","label_foreground","visual_background","cell_foreground","cell_background","kpi_value","icon"],"type":"string"}},"required":["field","id","rule","target"],"type":"object"},"type":"array"},"type":{"type":"string"}},"required":["type"],"type":"object"}],"description":"Strict point/scatter presentation and its governed result-frame channels.","properties":{"axes":{"items":{"additionalProperties":false,"description":"A renderer-neutral axis policy for Cartesian and point visuals.","properties":{"displayUnits":{"enum":["auto","none","thousands","millions","billions","trillions"],"type":"string"},"id":{"enum":["x","primary_y","secondary_y"],"type":"string"},"maximum":{"type":"number"},"minimum":{"type":"number"},"scale":{"enum":["automatic","linear","log"],"type":"string"},"tickDensity":{"enum":["automatic","sparse","normal","dense"],"type":"string"},"title":{"type":"string"},"unit":{"type":"string"},"zero":{"enum":["automatic","include","exclude"],"type":"string"}},"required":["id","scale","tickDensity","zero"],"type":"object"},"type":"array"},"brush":{"items":{"enum":["rectangle","lasso"],"type":"string"},"type":"array"},"color":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"colorScale":{"additionalProperties":false,"description":"A governed result-frame binding used by a point/scatter visual.","properties":{"kind":{"enum":["categorical","quantitative"],"type":"string"},"maximum":{"type":"number"},"minimum":{"type":"number"},"scheme":{"type":"string"}},"required":["kind"],"type":"object"},"eventAnnotations":{"items":{"additionalProperties":false,"description":"A governed event marker, restricted to the visual's horizontal axis.","properties":{"axis":{"enum":["x","primary_y","secondary_y"],"type":"string"},"description":{"type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"type":"string"},"tone":{"enum":["neutral","ink","success","warning","danger"],"type":"string"},"value":{"oneOf":[{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"kind":{"enum":["number"],"type":"string"},"value":{"type":"number"}},"required":["kind","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"kind":{"enum":["text"],"type":"string"},"value":{"type":"string"}},"required":["kind","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"kind":{"enum":["field"],"type":"string"},"reducer":{"enum":["first","last","minimum","maximum","mean","median"],"type":"string"}},"required":["field","kind","reducer"],"type":"object"}]}},"required":["axis","id","label","tone","value"],"type":"object"},"type":"array"},"identity":{"items":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"type":"array"},"label":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"labels":{"additionalProperties":false,"properties":{"density":{"enum":["hidden","automatic","dense","always"],"type":"string"},"maxCharacters":{"type":"integer"},"minimumSpacing":{"type":"integer"},"priority":{"items":{"enum":["selected","anomaly","threshold"],"type":"string"},"type":"array"},"tooltipFallback":{"type":"boolean"}},"required":["density"],"type":"object"},"legend":{"enum":["none","top","right","bottom","left"],"type":"string"},"overplot":{"additionalProperties":false,"description":"Overplotting policy for a point/scatter visual.","properties":{"largeMode":{"enum":["automatic","always","never"],"type":"string"},"largeThreshold":{"type":"integer"},"opacity":{"type":"number"},"strategy":{"enum":["show_all","opacity"],"type":"string"}},"required":["strategy"],"type":"object"},"referenceBands":{"items":{"additionalProperties":false,"description":"A governed interval between two horizontal or numeric reference values.","properties":{"axis":{"enum":["x","primary_y","secondary_y"],"type":"string"},"from":{"oneOf":[{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"kind":{"enum":["number"],"type":"string"},"value":{"type":"number"}},"required":["kind","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"kind":{"enum":["text"],"type":"string"},"value":{"type":"string"}},"required":["kind","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"kind":{"enum":["field"],"type":"string"},"reducer":{"enum":["first","last","minimum","maximum","mean","median"],"type":"string"}},"required":["field","kind","reducer"],"type":"object"}]},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"type":"string"},"to":{"oneOf":[{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"kind":{"enum":["number"],"type":"string"},"value":{"type":"number"}},"required":["kind","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"kind":{"enum":["text"],"type":"string"},"value":{"type":"string"}},"required":["kind","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"kind":{"enum":["field"],"type":"string"},"reducer":{"enum":["first","last","minimum","maximum","mean","median"],"type":"string"}},"required":["field","kind","reducer"],"type":"object"}]},"tone":{"enum":["neutral","ink","success","warning","danger"],"type":"string"}},"required":["axis","from","id","to","tone"],"type":"object"},"type":"array"},"referenceLines":{"items":{"additionalProperties":false,"description":"A governed horizontal or numeric reference marker.","properties":{"axis":{"enum":["x","primary_y","secondary_y"],"type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"type":"string"},"tone":{"enum":["neutral","ink","success","warning","danger"],"type":"string"},"value":{"oneOf":[{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"kind":{"enum":["number"],"type":"string"},"value":{"type":"number"}},"required":["kind","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"kind":{"enum":["text"],"type":"string"},"value":{"type":"string"}},"required":["kind","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"A scalar value used by a governed reference line, band, or annotation.","properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"kind":{"enum":["field"],"type":"string"},"reducer":{"enum":["first","last","minimum","maximum","mean","median"],"type":"string"}},"required":["field","kind","reducer"],"type":"object"}]}},"required":["axis","id","tone","value"],"type":"object"},"type":"array"},"series":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"size":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"sizeScale":{"additionalProperties":false,"description":"Pixel and optional domain bounds for a governed point size binding.","properties":{"maximum":{"type":"number"},"maximumPixels":{"type":"number"},"minimum":{"type":"number"},"minimumPixels":{"type":"number"}},"required":["maximumPixels","minimumPixels"],"type":"object"},"tooltip":{"items":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"type":"array"},"type":{"enum":["point"],"type":"string"},"x":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"y":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["identity","type","x","y"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"Renderer-neutral presentation policies; renderer options are intentionally absent.","properties":{"conditionalFormatting":{"items":{"additionalProperties":false,"description":"A closed, governed conditional-format binding over compiled result names.","properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"rule":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"high":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["gradient"],"type":"string"},"low":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"maximum":{"type":"number"},"minimum":{"type":"number"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"}},"required":["high","kind","low","maximum","minimum","nullStyle"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["rules"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"rules":{"items":{"additionalProperties":false,"description":"One ordered numeric comparison in a rules-based conditional format.","properties":{"operator":{"enum":["less_than","less_or_equal","greater_than","greater_or_equal","equal","not_equal"],"type":"string"},"style":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"value":{"type":"number"}},"required":["operator","style","value"],"type":"object"},"type":"array"}},"required":["defaultStyle","kind","nullStyle","rules"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["field"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"values":{"additionalProperties":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"type":"object"}},"required":["defaultStyle","kind","nullStyle","source","values"],"type":"object"}]},"target":{"enum":["mark_fill","mark_stroke","series_color","label_foreground","visual_background","cell_foreground","cell_background","kpi_value","icon"],"type":"string"}},"required":["field","id","rule","target"],"type":"object"},"type":"array"},"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"align":{"enum":["left","center","right"],"type":"string"},"centerLabel":{"type":"string"},"displayUnits":{"enum":["auto","none","thousands","millions","billions","trillions"],"type":"string"},"innerRadius":{"type":"number"},"labelPosition":{"enum":["automatic","inside","outside","top"],"type":"string"},"labels":{"additionalProperties":false,"properties":{"density":{"enum":["hidden","automatic","dense","always"],"type":"string"},"maxCharacters":{"type":"integer"},"minimumSpacing":{"type":"integer"},"priority":{"items":{"enum":["selected","anomaly","threshold"],"type":"string"},"type":"array"},"tooltipFallback":{"type":"boolean"}},"required":["density"],"type":"object"},"legend":{"enum":["none","top","right","bottom","left"],"type":"string"},"orientation":{"enum":["horizontal","vertical"],"type":"string"},"outerRadius":{"type":"number"},"rose":{"type":"boolean"},"sort":{"enum":["ascending","descending"],"type":"string"},"type":{"enum":["proportional"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"Renderer-neutral presentation policies; renderer options are intentionally absent.","properties":{"conditionalFormatting":{"items":{"additionalProperties":false,"description":"A closed, governed conditional-format binding over compiled result names.","properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"rule":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"high":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["gradient"],"type":"string"},"low":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"maximum":{"type":"number"},"minimum":{"type":"number"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"}},"required":["high","kind","low","maximum","minimum","nullStyle"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["rules"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"rules":{"items":{"additionalProperties":false,"description":"One ordered numeric comparison in a rules-based conditional format.","properties":{"operator":{"enum":["less_than","less_or_equal","greater_than","greater_or_equal","equal","not_equal"],"type":"string"},"style":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"value":{"type":"number"}},"required":["operator","style","value"],"type":"object"},"type":"array"}},"required":["defaultStyle","kind","nullStyle","rules"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["field"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"values":{"additionalProperties":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"type":"object"}},"required":["defaultStyle","kind","nullStyle","source","values"],"type":"object"}]},"target":{"enum":["mark_fill","mark_stroke","series_color","label_foreground","visual_background","cell_foreground","cell_background","kpi_value","icon"],"type":"string"}},"required":["field","id","rule","target"],"type":"object"},"type":"array"},"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"breadcrumb":{"type":"boolean"},"curveness":{"type":"number"},"focus":{"enum":["none","adjacency"],"type":"string"},"initialDepth":{"type":"integer"},"labels":{"additionalProperties":false,"properties":{"density":{"enum":["hidden","automatic","dense","always"],"type":"string"},"maxCharacters":{"type":"integer"},"minimumSpacing":{"type":"integer"},"priority":{"items":{"enum":["selected","anomaly","threshold"],"type":"string"},"type":"array"},"tooltipFallback":{"type":"boolean"}},"required":["density"],"type":"object"},"layout":{"enum":["standard","circular"],"type":"string"},"legend":{"enum":["none","top","right","bottom","left"],"type":"string"},"nodeGap":{"type":"number"},"orientation":{"enum":["horizontal","vertical"],"type":"string"},"roam":{"type":"boolean"},"type":{"enum":["hierarchy"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"Renderer-neutral presentation policies; renderer options are intentionally absent.","properties":{"conditionalFormatting":{"items":{"additionalProperties":false,"description":"A closed, governed conditional-format binding over compiled result names.","properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"rule":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"high":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["gradient"],"type":"string"},"low":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"maximum":{"type":"number"},"minimum":{"type":"number"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"}},"required":["high","kind","low","maximum","minimum","nullStyle"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["rules"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"rules":{"items":{"additionalProperties":false,"description":"One ordered numeric comparison in a rules-based conditional format.","properties":{"operator":{"enum":["less_than","less_or_equal","greater_than","greater_or_equal","equal","not_equal"],"type":"string"},"style":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"value":{"type":"number"}},"required":["operator","style","value"],"type":"object"},"type":"array"}},"required":["defaultStyle","kind","nullStyle","rules"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["field"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"values":{"additionalProperties":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"type":"object"}},"required":["defaultStyle","kind","nullStyle","source","values"],"type":"object"}]},"target":{"enum":["mark_fill","mark_stroke","series_color","label_foreground","visual_background","cell_foreground","cell_background","kpi_value","icon"],"type":"string"}},"required":["field","id","rule","target"],"type":"object"},"type":"array"},"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"area":{"type":"boolean"},"displayUnits":{"enum":["auto","none","thousands","millions","billions","trillions"],"type":"string"},"labels":{"additionalProperties":false,"properties":{"density":{"enum":["hidden","automatic","dense","always"],"type":"string"},"maxCharacters":{"type":"integer"},"minimumSpacing":{"type":"integer"},"priority":{"items":{"enum":["selected","anomaly","threshold"],"type":"string"},"type":"array"},"tooltipFallback":{"type":"boolean"}},"required":["density"],"type":"object"},"legend":{"enum":["none","top","right","bottom","left"],"type":"string"},"maximum":{"type":"number"},"minimum":{"type":"number"},"progressWidth":{"type":"number"},"showPointer":{"type":"boolean"},"target":{"type":"number"},"thresholds":{"items":{"additionalProperties":false,"properties":{"tone":{"enum":["neutral","ink","success","warning","danger"],"type":"string"},"value":{"type":"number"}},"required":["tone","value"],"type":"object"},"type":"array"},"type":{"enum":["polar"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"Renderer-neutral presentation policies; renderer options are intentionally absent.","properties":{"conditionalFormatting":{"items":{"additionalProperties":false,"description":"A closed, governed conditional-format binding over compiled result names.","properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"rule":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"high":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["gradient"],"type":"string"},"low":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"maximum":{"type":"number"},"minimum":{"type":"number"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"}},"required":["high","kind","low","maximum","minimum","nullStyle"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["rules"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"rules":{"items":{"additionalProperties":false,"description":"One ordered numeric comparison in a rules-based conditional format.","properties":{"operator":{"enum":["less_than","less_or_equal","greater_than","greater_or_equal","equal","not_equal"],"type":"string"},"style":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"value":{"type":"number"}},"required":["operator","style","value"],"type":"object"},"type":"array"}},"required":["defaultStyle","kind","nullStyle","rules"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["field"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"values":{"additionalProperties":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"type":"object"}},"required":["defaultStyle","kind","nullStyle","source","values"],"type":"object"}]},"target":{"enum":["mark_fill","mark_stroke","series_color","label_foreground","visual_background","cell_foreground","cell_background","kpi_value","icon"],"type":"string"}},"required":["field","id","rule","target"],"type":"object"},"type":"array"},"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"basemap":{"type":"string"},"camera":{"additionalProperties":false,"properties":{"center":{"items":{"type":"number"},"type":"array"},"maximumZoom":{"type":"number"},"minimumZoom":{"type":"number"},"mode":{"enum":["fit_data","fixed","preserve"],"type":"string"},"padding":{"type":"integer"},"zoom":{"type":"number"}},"type":"object"},"controls":{"additionalProperties":false,"properties":{"compass":{"type":"boolean"},"reset":{"type":"boolean"},"zoom":{"type":"boolean"}},"type":"object"},"labelDensity":{"enum":["hidden","normal","dense"],"type":"string"},"labels":{"additionalProperties":false,"properties":{"density":{"enum":["hidden","automatic","dense","always"],"type":"string"},"maxCharacters":{"type":"integer"},"minimumSpacing":{"type":"integer"},"priority":{"items":{"enum":["selected","anomaly","threshold"],"type":"string"},"type":"array"},"tooltipFallback":{"type":"boolean"}},"required":["density"],"type":"object"},"layers":{"items":{"oneOf":[{"additionalProperties":false,"allOf":[{"allOf":[{"properties":{"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"maximumZoom":{"type":"number"},"minimumZoom":{"type":"number"},"position":{"enum":["below_labels","above_labels"],"type":"string"},"tooltip":{"items":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"type":"array"}},"required":["id"],"type":"object"}],"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"category":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"cluster":{"additionalProperties":false,"properties":{"enabled":{"type":"boolean"},"maximumZoom":{"type":"integer"},"minimumPoints":{"type":"integer"},"radius":{"type":"integer"},"showCount":{"type":"boolean"}},"type":"object"},"color":{"additionalProperties":false,"properties":{"domainMaximum":{"type":"number"},"domainMidpoint":{"type":"number"},"domainMinimum":{"type":"number"},"kind":{"enum":["sequential","diverging","categorical"],"type":"string"},"nullColor":{"type":"string"},"palette":{"type":"string"},"reverse":{"type":"boolean"}},"type":"object"},"kind":{"enum":["point"],"type":"string"},"latitude":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"longitude":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"opacity":{"type":"number"},"size":{"additionalProperties":false,"properties":{"domainMaximum":{"type":"number"},"domainMinimum":{"type":"number"},"maximumRadius":{"type":"number"},"minimumRadius":{"type":"number"}},"type":"object"},"stroke":{"additionalProperties":false,"properties":{"color":{"type":"string"},"opacity":{"type":"number"},"width":{"type":"number"}},"type":"object"},"value":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["kind","latitude","longitude"],"type":"object"},{"additionalProperties":false,"allOf":[{"allOf":[{"properties":{"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"maximumZoom":{"type":"number"},"minimumZoom":{"type":"number"},"position":{"enum":["below_labels","above_labels"],"type":"string"},"tooltip":{"items":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"type":"array"}},"required":["id"],"type":"object"}],"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"category":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"color":{"additionalProperties":false,"properties":{"domainMaximum":{"type":"number"},"domainMidpoint":{"type":"number"},"domainMinimum":{"type":"number"},"kind":{"enum":["sequential","diverging","categorical"],"type":"string"},"nullColor":{"type":"string"},"palette":{"type":"string"},"reverse":{"type":"boolean"}},"type":"object"},"geometryAsset":{"type":"string"},"join":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"kind":{"enum":["choropleth"],"type":"string"},"opacity":{"type":"number"},"stroke":{"additionalProperties":false,"properties":{"color":{"type":"string"},"opacity":{"type":"number"},"width":{"type":"number"}},"type":"object"},"value":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["geometryAsset","join","kind"],"type":"object"},{"additionalProperties":false,"allOf":[{"allOf":[{"properties":{"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"maximumZoom":{"type":"number"},"minimumZoom":{"type":"number"},"position":{"enum":["below_labels","above_labels"],"type":"string"},"tooltip":{"items":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"type":"array"}},"required":["id"],"type":"object"}],"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"color":{"additionalProperties":false,"properties":{"domainMaximum":{"type":"number"},"domainMidpoint":{"type":"number"},"domainMinimum":{"type":"number"},"kind":{"enum":["sequential","diverging","categorical"],"type":"string"},"nullColor":{"type":"string"},"palette":{"type":"string"},"reverse":{"type":"boolean"}},"type":"object"},"geometryAsset":{"type":"string"},"kind":{"enum":["reference"],"type":"string"},"opacity":{"type":"number"},"stroke":{"additionalProperties":false,"properties":{"color":{"type":"string"},"opacity":{"type":"number"},"width":{"type":"number"}},"type":"object"}},"required":["geometryAsset","kind"],"type":"object"},{"additionalProperties":false,"allOf":[{"allOf":[{"properties":{"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"maximumZoom":{"type":"number"},"minimumZoom":{"type":"number"},"position":{"enum":["below_labels","above_labels"],"type":"string"},"tooltip":{"items":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"type":"array"}},"required":["id"],"type":"object"}],"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"color":{"additionalProperties":false,"properties":{"domainMaximum":{"type":"number"},"domainMidpoint":{"type":"number"},"domainMinimum":{"type":"number"},"kind":{"enum":["sequential","diverging","categorical"],"type":"string"},"nullColor":{"type":"string"},"palette":{"type":"string"},"reverse":{"type":"boolean"}},"type":"object"},"heat":{"additionalProperties":false,"properties":{"intensity":{"type":"number"},"radius":{"type":"number"}},"type":"object"},"kind":{"enum":["heat"],"type":"string"},"latitude":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"longitude":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"opacity":{"type":"number"},"value":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["kind","latitude","longitude"],"type":"object"},{"additionalProperties":false,"allOf":[{"allOf":[{"properties":{"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"maximumZoom":{"type":"number"},"minimumZoom":{"type":"number"},"position":{"enum":["below_labels","above_labels"],"type":"string"},"tooltip":{"items":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"type":"array"}},"required":["id"],"type":"object"}],"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"color":{"additionalProperties":false,"properties":{"domainMaximum":{"type":"number"},"domainMidpoint":{"type":"number"},"domainMinimum":{"type":"number"},"kind":{"enum":["sequential","diverging","categorical"],"type":"string"},"nullColor":{"type":"string"},"palette":{"type":"string"},"reverse":{"type":"boolean"}},"type":"object"},"heat":{"additionalProperties":false,"properties":{"intensity":{"type":"number"},"radius":{"type":"number"}},"type":"object"},"kind":{"enum":["density"],"type":"string"},"latitude":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"longitude":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"opacity":{"type":"number"},"value":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["kind","latitude","longitude"],"type":"object"},{"additionalProperties":false,"allOf":[{"allOf":[{"properties":{"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"label":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"maximumZoom":{"type":"number"},"minimumZoom":{"type":"number"},"position":{"enum":["below_labels","above_labels"],"type":"string"},"tooltip":{"items":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"type":"array"}},"required":["id"],"type":"object"}],"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"category":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"color":{"additionalProperties":false,"properties":{"domainMaximum":{"type":"number"},"domainMidpoint":{"type":"number"},"domainMinimum":{"type":"number"},"kind":{"enum":["sequential","diverging","categorical"],"type":"string"},"nullColor":{"type":"string"},"palette":{"type":"string"},"reverse":{"type":"boolean"}},"type":"object"},"kind":{"enum":["path"],"type":"string"},"latitude":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"line":{"additionalProperties":false,"properties":{"curvature":{"type":"number"},"width":{"type":"number"}},"type":"object"},"longitude":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"opacity":{"type":"number"},"order":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"path":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"stroke":{"additionalProperties":false,"properties":{"color":{"type":"string"},"opacity":{"type":"number"},"width":{"type":"number"}},"type":"object"},"value":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["kind","latitude","longitude","order","path"],"type":"object"}]},"type":"array"},"roam":{"type":"boolean"},"theme":{"enum":["auto","light","dark"],"type":"string"},"type":{"enum":["geographic"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"Renderer-neutral presentation policies; renderer options are intentionally absent.","properties":{"conditionalFormatting":{"items":{"additionalProperties":false,"description":"A closed, governed conditional-format binding over compiled result names.","properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"rule":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"high":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["gradient"],"type":"string"},"low":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"maximum":{"type":"number"},"minimum":{"type":"number"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"}},"required":["high","kind","low","maximum","minimum","nullStyle"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["rules"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"rules":{"items":{"additionalProperties":false,"description":"One ordered numeric comparison in a rules-based conditional format.","properties":{"operator":{"enum":["less_than","less_or_equal","greater_than","greater_or_equal","equal","not_equal"],"type":"string"},"style":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"value":{"type":"number"}},"required":["operator","style","value"],"type":"object"},"type":"array"}},"required":["defaultStyle","kind","nullStyle","rules"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["field"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"values":{"additionalProperties":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"type":"object"}},"required":["defaultStyle","kind","nullStyle","source","values"],"type":"object"}]},"target":{"enum":["mark_fill","mark_stroke","series_color","label_foreground","visual_background","cell_foreground","cell_background","kpi_value","icon"],"type":"string"}},"required":["field","id","rule","target"],"type":"object"},"type":"array"},"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"rowHeight":{"type":"integer"},"showHeader":{"type":"boolean"},"striped":{"type":"boolean"},"type":{"enum":["table"],"type":"string"}},"required":["rowHeight","showHeader","striped","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"description":"Renderer-neutral presentation policies; renderer options are intentionally absent.","properties":{"conditionalFormatting":{"items":{"additionalProperties":false,"description":"A closed, governed conditional-format binding over compiled result names.","properties":{"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"id":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"rule":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"high":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["gradient"],"type":"string"},"low":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"maximum":{"type":"number"},"minimum":{"type":"number"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"}},"required":["high","kind","low","maximum","minimum","nullStyle"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["rules"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"rules":{"items":{"additionalProperties":false,"description":"One ordered numeric comparison in a rules-based conditional format.","properties":{"operator":{"enum":["less_than","less_or_equal","greater_than","greater_or_equal","equal","not_equal"],"type":"string"},"style":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"value":{"type":"number"}},"required":["operator","style","value"],"type":"object"},"type":"array"}},"required":["defaultStyle","kind","nullStyle","rules"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"kind":{"type":"string"}},"required":["kind"],"type":"object"}],"properties":{"defaultStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"kind":{"enum":["field"],"type":"string"},"nullStyle":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"source":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"values":{"additionalProperties":{"additionalProperties":false,"description":"A renderer-neutral color or icon intent applied by a conditional rule.","properties":{"color":{"enum":["accent","neutral","ink","success","warning","danger","data_1","data_2","data_3","data_4","data_5","data_6","data_7","data_8"],"type":"string"},"icon":{"enum":["circle","square","diamond","triangle_up","triangle_down","arrow_up","arrow_down","warning"],"type":"string"}},"type":"object"},"type":"object"}},"required":["defaultStyle","kind","nullStyle","source","values"],"type":"object"}]},"target":{"enum":["mark_fill","mark_stroke","series_color","label_foreground","visual_background","cell_foreground","cell_background","kpi_value","icon"],"type":"string"}},"required":["field","id","rule","target"],"type":"object"},"type":"array"},"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"comparison":{"additionalProperties":false,"description":"A governed numeric result used for a KPI comparison or goal.","properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"label":{"type":"string"},"reducer":{"enum":["first","last","minimum","maximum","mean","median"],"type":"string"}},"required":["dataset","field","label"],"type":"object"},"delta":{"enum":["absolute","relative"],"type":"string"},"displayUnits":{"enum":["auto","none","thousands","millions","billions","trillions"],"type":"string"},"favorableDirection":{"enum":["increase","decrease","neutral"],"type":"string"},"goal":{"additionalProperties":false,"description":"A governed numeric result used for a KPI comparison or goal.","properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"label":{"type":"string"},"reducer":{"enum":["first","last","minimum","maximum","mean","median"],"type":"string"}},"required":["dataset","field","label"],"type":"object"},"missingComparison":{"enum":["show_unavailable","hide"],"type":"string"},"mode":{"enum":["compact","bullet","progress"],"type":"string"},"note":{"type":"string"},"ranges":{"items":{"additionalProperties":false,"properties":{"label":{"type":"string"},"maximum":{"type":"number"},"minimum":{"type":"number"},"tone":{"enum":["neutral","ink","success","warning","danger"],"type":"string"}},"required":["label","tone"],"type":"object"},"type":"array"},"thresholds":{"items":{"additionalProperties":false,"properties":{"tone":{"enum":["neutral","ink","success","warning","danger"],"type":"string"},"value":{"type":"number"}},"required":["tone","value"],"type":"object"},"type":"array"},"tone":{"enum":["neutral","ink","success","warning","danger"],"type":"string"},"trend":{"additionalProperties":false,"description":"A governed result frame used for a KPI sparkline trend.","properties":{"category":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"value":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["category","dataset","value"],"type":"object"},"type":{"enum":["kpi"],"type":"string"}},"required":["type"],"type":"object"}]},"query":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"dimensions":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic dimension reference. Physical dataset fields are not accepted.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"dimension":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"grain":{"enum":["second","minute","hour","day","week","month","quarter","year"],"type":"string"}},"required":["dimension"],"type":"object"}]},"type":"array"},"limit":{"type":"integer"},"metrics":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic metric reference. Metric expressions are never part of this type.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"metric":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["metric"],"type":"object"}]},"type":"array"},"sort":{"items":{"additionalProperties":false,"properties":{"direction":{"enum":["asc","desc"],"type":"string"},"field":{"description":"A compiled result-frame field name, never a source member or expression.","minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["direction","field"],"type":"object"},"type":"array"},"type":{"enum":["aggregate"],"type":"string"}},"required":["dimensions","metrics","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"dataset":{"maxLength":128,"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_-]*$","type":"string"},"fields":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A physical field reference is only valid for a records query root dataset.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"field":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["field"],"type":"object"}]},"type":"array"},"limit":{"type":"integer"},"sort":{"items":{"additionalProperties":false,"properties":{"direction":{"enum":["asc","desc"],"type":"string"},"field":{"description":"A compiled result-frame field name, never a source member or expression.","minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["direction","field"],"type":"object"},"type":"array"},"type":{"enum":["records"],"type":"string"}},"required":["dataset","fields","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"columns":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic dimension reference. Physical dataset fields are not accepted.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"dimension":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"grain":{"enum":["second","minute","hour","day","week","month","quarter","year"],"type":"string"}},"required":["dimension"],"type":"object"}]},"type":"array"},"metrics":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic metric reference. Metric expressions are never part of this type.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"metric":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["metric"],"type":"object"}]},"type":"array"},"rows":{"items":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic dimension reference. Physical dataset fields are not accepted.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"dimension":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"grain":{"enum":["second","minute","hour","day","week","month","quarter","year"],"type":"string"}},"required":["dimension"],"type":"object"}]},"type":"array"},"sort":{"items":{"additionalProperties":false,"properties":{"direction":{"enum":["asc","desc"],"type":"string"},"field":{"description":"A compiled result-frame field name, never a source member or expression.","minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["direction","field"],"type":"object"},"type":"array"},"totals":{"additionalProperties":false,"properties":{"columns":{"type":"boolean"},"grand":{"type":"boolean"},"rows":{"type":"boolean"}},"type":"object"},"type":{"enum":["pivot"],"type":"string"},"window":{"additionalProperties":false,"properties":{"limit":{"type":"integer"},"offset":{"type":"integer"}},"required":["limit"],"type":"object"}},"required":["columns","metrics","rows","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"approximation":{"enum":["exact","approximate"],"type":"string"},"bins":{"type":"integer"},"domain":{"additionalProperties":false,"properties":{"maximum":{"type":"number"},"minimum":{"type":"number"}},"type":"object"},"field":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic metric reference. Metric expressions are never part of this type.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"metric":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["metric"],"type":"object"}]},"nullPolicy":{"enum":["omit","include"],"type":"string"},"type":{"enum":["histogram"],"type":"string"}},"required":["approximation","bins","field","nullPolicy","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"approximation":{"enum":["exact","approximate"],"type":"string"},"field":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic metric reference. Metric expressions are never part of this type.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"metric":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"}},"required":["metric"],"type":"object"}]},"group":{"oneOf":[{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},{"additionalProperties":false,"description":"A semantic dimension reference. Physical dataset fields are not accepted.","properties":{"alias":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"dimension":{"minLength":1,"pattern":"^[A-Za-z_][A-Za-z0-9_]*$","type":"string"},"grain":{"enum":["second","minute","hour","day","week","month","quarter","year"],"type":"string"}},"required":["dimension"],"type":"object"}]},"limit":{"type":"integer"},"outliers":{"enum":["omit","include"],"type":"string"},"quantiles":{"items":{"type":"number"},"type":"array"},"type":{"enum":["distribution"],"type":"string"},"whiskers":{"additionalProperties":false,"properties":{"lower":{"type":"number"},"upper":{"type":"number"}},"required":["lower","upper"],"type":"object"}},"required":["approximation","field","outliers","quantiles","type"],"type":"object"}]},"subtitle":{"type":"string"},"title":{"type":"string"},"type":{"enum":["line","area","bar","column","pie","donut","scatter","funnel","treemap","gauge","heatmap","sankey","graph","map","candlestick","boxplot","combo","waterfall","histogram","radar","tree","sunburst","kpi","table","matrix","pivot"],"type":"string"}},"required":["presentation","query","type"],"type":"object"}},"required":["semanticModelId","visual"],"type":"object"}`
 	QueryVisualResultSchemaJSON = `{"additionalProperties":false,"properties":{"completeness":{"additionalProperties":false,"properties":{"limit":{"type":"integer"},"returnedRows":{"type":"integer"},"status":{"enum":["complete","limit_reached","empty"],"type":"string"}},"required":["limit","returnedRows","status"],"type":"object"},"datasetId":{"type":"string"},"diagnostics":{"items":{"additionalProperties":false,"properties":{"code":{"type":"string"},"fieldId":{"type":"string"},"message":{"type":"string"},"severity":{"enum":["info","warning","error"],"type":"string"}},"required":["code","message","severity"],"type":"object"},"type":"array"},"fields":{"items":{"additionalProperties":false,"properties":{"alias":{"type":"string"},"dataType":{"type":"string"},"fieldId":{"type":"string"},"format":{"type":"string"},"label":{"type":"string"},"role":{"enum":["dimension","series","metric","table_field","table_row"],"type":"string"},"unit":{"type":"string"}},"required":["fieldId","label","role"],"type":"object"},"type":"array"},"filters":{"items":{"additionalProperties":false,"properties":{"expression":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["unfiltered"],"type":"string"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"operator":{"enum":["isNull","isNotNull","in","notIn","equals","notEquals","contains","notContains","startsWith","endsWith","greaterThan","greaterThanOrEqual","lessThan","lessThanOrEqual"],"type":"string"},"type":{"enum":["nullCheck"],"type":"string"}},"required":["operator","type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"operator":{"enum":["isNull","isNotNull","in","notIn","equals","notEquals","contains","notContains","startsWith","endsWith","greaterThan","greaterThanOrEqual","lessThan","lessThanOrEqual"],"type":"string"},"type":{"enum":["set"],"type":"string"},"values":{"items":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["string"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["boolean"],"type":"string"},"value":{"type":"boolean"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["integer"],"type":"string"},"value":{"description":"Precision-safe integer representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["decimal"],"type":"string"},"value":{"description":"Canonical decimal representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["date"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["timestamp"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"}]},"type":"array"}},"required":["operator","type","values"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"operator":{"enum":["isNull","isNotNull","in","notIn","equals","notEquals","contains","notContains","startsWith","endsWith","greaterThan","greaterThanOrEqual","lessThan","lessThanOrEqual"],"type":"string"},"type":{"enum":["comparison"],"type":"string"},"value":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["string"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["boolean"],"type":"string"},"value":{"type":"boolean"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["integer"],"type":"string"},"value":{"description":"Precision-safe integer representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["decimal"],"type":"string"},"value":{"description":"Canonical decimal representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["date"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["timestamp"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"}]}},"required":["operator","type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"lower":{"additionalProperties":false,"properties":{"inclusive":{"type":"boolean"},"value":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["string"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["boolean"],"type":"string"},"value":{"type":"boolean"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["integer"],"type":"string"},"value":{"description":"Precision-safe integer representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["decimal"],"type":"string"},"value":{"description":"Canonical decimal representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["date"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["timestamp"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"}]}},"required":["inclusive","value"],"type":"object"},"type":{"enum":["range"],"type":"string"},"upper":{"additionalProperties":false,"properties":{"inclusive":{"type":"boolean"},"value":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["string"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["boolean"],"type":"string"},"value":{"type":"boolean"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["integer"],"type":"string"},"value":{"description":"Precision-safe integer representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["decimal"],"type":"string"},"value":{"description":"Canonical decimal representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["date"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["timestamp"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"}]}},"required":["inclusive","value"],"type":"object"}},"required":["type"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"anchor":{"enum":["currentTime","firstAvailable","lastAvailable","fixed"],"type":"string"},"anchorValue":{"oneOf":[{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["string"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["boolean"],"type":"string"},"value":{"type":"boolean"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["integer"],"type":"string"},"value":{"description":"Precision-safe integer representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["decimal"],"type":"string"},"value":{"description":"Canonical decimal representation.","type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["date"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},{"additionalProperties":false,"allOf":[{"properties":{"type":{"type":"string"}},"required":["type"],"type":"object"}],"properties":{"type":{"enum":["timestamp"],"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"}]},"count":{"type":"integer"},"direction":{"enum":["previous","current","next"],"type":"string"},"includeCurrent":{"type":"boolean"},"type":{"enum":["relativePeriod"],"type":"string"},"unit":{"enum":["minute","hour","day","week","month","quarter","year"],"type":"string"}},"required":["anchor","count","direction","includeCurrent","type","unit"],"type":"object"}]},"fieldId":{"description":"Filtered semantic field.","type":"string"},"path":{"description":"Zero-based nested group path. Omitted for a root-level filter.","items":{"type":"integer"},"type":"array"},"resolvedDatasetId":{"description":"Semantic dataset selected to resolve a conformed dimension.","type":"string"}},"required":["expression","fieldId"],"type":"object"},"type":"array"},"freshness":{"additionalProperties":false,"properties":{"lastSuccessfulRefreshAt":{"description":"Completion time of the last successful publish or refresh. This is not an upstream event-time watermark.","type":"string"},"servingStateId":{"description":"Serving state that produced the successful data version.","type":"string"},"snapshotId":{"description":"DuckLake snapshot ID recorded for the successful data version.","type":"string"},"source":{"description":"Whether the data version came from publish or refresh.","enum":["publish","refresh"],"type":"string"},"status":{"description":"current when the successful data version belongs to the queried serving snapshot; stale otherwise.","enum":["current","stale"],"type":"string"}},"required":["lastSuccessfulRefreshAt","servingStateId","snapshotId","source","status"],"type":"object"},"id":{"type":"string"},"ok":{"type":"boolean"},"queryId":{"type":"string"},"semanticModelRef":{"additionalProperties":false,"properties":{"id":{"maxLength":256,"minLength":1,"pattern":"^[A-Za-z0-9][A-Za-z0-9_.:-]*$","type":"string"},"kind":{"enum":["project","connection","source","model","semantic_model","pipeline","dashboard"],"type":"string"}},"required":["id","kind"],"type":"object"},"servingSnapshot":{"type":"string"},"signal":{"type":"string"},"status":{"additionalProperties":false,"properties":{"kind":{"enum":["idle","loading","ready","no_data","partial","error"],"type":"string"},"message":{"type":"string"}},"required":["kind"],"type":"object"},"summary":{"type":"string"},"title":{"type":"string"},"type":{"type":"string"}},"required":["completeness","datasetId","diagnostics","fields","filters","id","ok","queryId","semanticModelRef","servingSnapshot","signal","status","summary","title","type"],"type":"object"}`
 )
